backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-dotenv==1.0.0
python-multipart==0.0.9
zhipuai==2.0.1